Merge latest master from upstream

+++ insights/client/client.py
@@ -8,9 +8,8 @@ import os
 import time
 import shutil
 import six
+import atexit
 
-from .auto_config import (_try_satellite6_configuration,
-                          _try_satellite5_configuration)
 from .utilities import (generate_machine_id,
                         write_to_disk,
                         write_registered_file,
@@ -24,9 +23,9 @@ from .connection import InsightsConnection
 from .archive import InsightsArchive
 from .support import registration_check
 from .constants import InsightsConstants as constants
+from .schedule import get_scheduler
 
 LOG_FORMAT = ("%(asctime)s %(levelname)8s %(name)s %(message)s")
-INSIGHTS_CONNECTION = None
 logger = logging.getLogger(__name__)
 
 
@@ -78,6 +77,8 @@ def configure_level(config):
 
     net_debug_level = logging.INFO if config.net_debug else logging.ERROR
     logging.getLogger('network').setLevel(net_debug_level)
+    if not config.verbose:
+        logging.getLogger('insights.core.dr').setLevel(logging.WARNING)
 
 
 def set_up_logging(config):
@@ -88,79 +89,8 @@ def set_up_logging(config):
         logger.debug("Logging initialized")
 
 
-def test_connection(config):
-    """
-    Test the connection
-    """
-    pconn = get_connection(config)
-    return pconn.test_connection()
-
-
-def _is_client_registered(config):
-
-    # If the client is running in container mode, bypass this stuff
-    if config.analyze_container:
-        return 'Client running in container/image mode. Bypassing registration check', False
-
-    # All other cases
-    msg_notyet = 'This machine has not yet been registered.'
-    msg_unreg = 'This machine has been unregistered.'
-    msg_doreg = 'Use --register to register this machine.'
-    msg_rereg = 'Use --register if you would like to re-register this machine.'
-
-    # check reg status w/ API
-    reg_check = registration_check(get_connection(config))
-    if not reg_check['status']:
-        # not registered
-        if reg_check['unreg_date']:
-            # system has been unregistered from the UI
-            msg = '\n'.join([msg_unreg, msg_rereg])
-            write_unregistered_file(reg_check['unreg_date'])
-            return msg, False
-        else:
-            # no record of system in remote
-            msg = '\n'.join([msg_notyet, msg_doreg])
-            # clear any local records
-            delete_registered_file()
-            delete_unregistered_file()
-            return msg, False
-    else:
-        # API confirms reg
-        write_registered_file()
-        return '', True
-
-
-def try_register(config):
-
-    # if we are running an image analysis then dont register
-    if config.analyze_container:
-        logger.info("Running client in Container mode. Bypassing registration.")
-        return
-
-    # check reg status with API
-    reg_check = registration_check(get_connection(config))
-    if reg_check['status']:
-        logger.info('This host has already been registered.')
-        # regenerate the .registered file
-        write_registered_file()
-        return True
-    if reg_check['unreachable']:
-        logger.error(reg_check['messages'][1])
-        return None
-    message, hostname, group, display_name = register(config)
-    if config.display_name is None and config.group is None:
-        logger.info('Successfully registered host %s', hostname)
-    elif config.display_name is None:
-        logger.info('Successfully registered host %s in group %s', hostname, group)
-    else:
-        logger.info('Successfully registered host %s as %s in group %s',
-                    hostname, display_name, group)
-    if message:
-        logger.info(message)
-    return reg_check, message, hostname, group, display_name
-
-
-def register(config):
+# -LEGACY-
+def register(config, pconn):
     """
     Do registration using basic auth
     """
@@ -171,60 +101,156 @@ def register(config):
     if not username and not password and not auto_config and authmethod == 'BASIC':
         logger.debug('Username and password must be defined in configuration file with BASIC authentication method.')
         return False
-    pconn = get_connection(config)
     return pconn.register()
 
 
-def handle_registration(config):
-    """
-        returns (json): {'success': bool,
-                        'machine-id': uuid from API,
-                        'response': response from API,
-                        'code': http code}
-    """
+# -LEGACY-
+def _legacy_handle_registration(config, pconn):
+    '''
+    Handle the registration process
+    Returns:
+        True - machine is registered
+        False - machine is unregistered
+        None - could not reach the API
+    '''
+    logger.debug('Trying registration.')
     # force-reregister -- remove machine-id files and registration files
     # before trying to register again
-    new = False
     if config.reregister:
-        logger.debug('Re-register set, forcing registration.')
-        new = True
-        config.register = True
         delete_registered_file()
         delete_unregistered_file()
         write_to_disk(constants.machine_id_file, delete=True)
-    logger.debug('Machine-id: %s', generate_machine_id(new))
+        logger.debug('Re-register set, forcing registration.')
 
-    logger.debug('Trying registration.')
-    registration = try_register()
-    if registration is None:
+    logger.debug('Machine-id: %s', generate_machine_id(new=config.reregister))
+
+    # check registration with API
+    check = get_registration_status(config, pconn)
+
+    for m in check['messages']:
+        logger.debug(m)
+
+    if check['unreachable']:
+        # Run connection test and exit
         return None
-    msg, is_registered = _is_client_registered()
 
-    return {'success': is_registered,
-            'machine-id': generate_machine_id(),
-            'registration': registration}
+    if check['status']:
+        # registered in API, resync files
+        if config.register:
+            logger.info('This host has already been registered.')
+        write_registered_file()
+        return True
+
+    if config.register:
+        if config.group:
+            logger.info('Registering with a group is currently unavailable.')
+        # register if specified
+        message, hostname, group, display_name = register(config, pconn)
+        if not hostname:
+            # API could not be reached, run connection test and exit
+            logger.error(message)
+            return None
+        if config.display_name is None and config.group is None:
+            logger.info('Successfully registered host %s', hostname)
+        elif config.display_name is None:
+            logger.info('Successfully registered host %s in group %s',
+                        hostname, group)
+        else:
+            logger.info('Successfully registered host %s as %s in group %s',
+                        hostname, display_name, group)
+        if message:
+            logger.info(message)
+        write_registered_file()
+        return True
+    else:
+        # unregistered in API, resync files
+        write_unregistered_file(date=check['unreg_date'])
+        # print messaging and exit
+        if check['unreg_date']:
+            # registered and then unregistered
+            logger.info('This machine has been unregistered. '
+                        'Use --register if you would like to '
+                        're-register this machine.')
+        else:
+            # not yet registered
+            logger.info('This machine has not yet been registered.'
+                        'Use --register to register this machine.')
+        return False
 
 
-def get_registration_status(config):
-    return registration_check(get_connection(config))
+def handle_registration(config, pconn):
+    '''
+    Does nothing on the platform. Will be deleted eventually.
+    '''
+    if config.legacy_upload:
+        return _legacy_handle_registration(config, pconn)
 
 
-def handle_unregistration(config):
+def get_registration_status(config, pconn):
+    '''
+        Handle the registration process
+        Returns:
+            True - machine is registered
+            False - machine is unregistered
+            None - could not reach the API
+    '''
+    return registration_check(pconn)
+
+
+# -LEGACY-
+def _legacy_handle_unregistration(config, pconn):
     """
         returns (bool): True success, False failure
     """
-    pconn = get_connection(config)
-    return pconn.unregister()
+    check = get_registration_status(config, pconn)
+
+    for m in check['messages']:
+        logger.debug(m)
+
+    if check['unreachable']:
+        # Run connection test and exit
+        return None
+
+    if check['status']:
+        unreg = pconn.unregister()
+    else:
+        unreg = True
+        logger.info('This system is already unregistered.')
+    if unreg:
+        # only set if unreg was successful
+        write_unregistered_file()
+        get_scheduler(config).remove_scheduling()
+    return unreg
+
+
+def handle_unregistration(config, pconn):
+    """
+    Returns:
+        True - machine was successfully unregistered
+        False - machine could not be unregistered
+        None - could not reach the API
+    """
+    if config.legacy_upload:
+        return _legacy_handle_unregistration(config, pconn)
+
+    unreg = pconn.unregister()
+    if unreg:
+        # only set if unreg was successful
+        write_unregistered_file()
+    return unreg
 
 
 def get_machine_id():
     return generate_machine_id()
 
 
-def update_rules(config):
-    pconn = get_connection(config)
+def update_rules(config, pconn):
+    if not pconn:
+        raise ValueError('ERROR: Cannot update rules in --offline mode. '
+                         'Disable auto_update in config file.')
+
     pc = InsightsUploadConf(config, conn=pconn)
-    return pc.get_conf(True, {})
+    return pc.get_conf_update()
 
 
 def get_branch_info(config):
@@ -232,27 +258,16 @@ def get_branch_info(config):
     Get branch info for a system
     returns (dict): {'remote_branch': -1, 'remote_leaf': -1}
     """
-    branch_info = constants.default_branch_info
-
     # in the case we are running on offline mode
     # or we are analyzing a running container/image
     # or tar file, mountpoint, simply return the default branch info
     if (config.offline or
             config.analyze_container):
-        return branch_info
-
-    # otherwise continue reaching out to obtain branch info
-    try:
-        pconn = get_connection(config)
-        branch_info = pconn.branch_info()
-    except LookupError:
-        logger.debug("There was an error obtaining branch information.")
-        logger.debug("Assuming default branch information %s" % branch_info)
-    logger.debug("Obtained branch information: %s" % branch_info)
-    return branch_info
+        return constants.default_branch_info
+    return config.branch_info
 
 
-def collect(config):
+def collect(config, pconn):
     """
     All the heavy lifting done here
     """
@@ -273,10 +288,10 @@ def collect(config):
                         os.path.basename(config.analyze_mountpoint))[0],
                   'location': config.analyze_mountpoint}
 
-    # container
-    elif config.analyze_container or config.analyze_image_id:
-        logger.debug("Client running in container/image mode.")
-        logger.debug("Scanning for matching container/image.")
+    # image
+    elif config.analyze_image_id:
+        logger.debug("Client running in image mode.")
+        logger.debug("Scanning for matching image.")
 
         from .containers import get_targets
         targets = get_targets(config)
@@ -284,36 +299,20 @@ def collect(config):
             sys.exit(constants.sig_kill_bad)
         target = targets[0]
 
-    # the host
+    # host, or inside container
     else:
-        logger.debug("Host selected as scanning target.")
+        if config.analyze_container:
+            logger.debug('Client running in container mode.')
+        else:
+            logger.debug("Host selected as scanning target.")
         target = constants.default_target
 
     branch_info = get_branch_info(config)
     pc = InsightsUploadConf(config)
     tar_file = None
 
-    # load config from stdin/file if specified
-    try:
-        stdin_config = {}
-        if config.from_file:
-            with open(config.from_file, 'r') as f:
-                stdin_config = json.load(f)
-        elif config.from_stdin:
-            stdin_config = json.load(sys.stdin)
-        if ((config.from_file or config.from_stdin) and
-            ('uploader.json' not in stdin_config or
-             'sig' not in stdin_config)):
-            raise ValueError
-        if ((config.from_file or config.from_stdin) and
-                'branch_info' in stdin_config and stdin_config['branch_info'] is not None):
-            branch_info = stdin_config['branch_info']
-    except:
-        logger.error('ERROR: Invalid config for %s! Exiting...',
-                     ('--from-file' if config.from_file else '--from-stdin'))
-        return False
-
-    collection_rules, rm_conf = pc.get_conf(False, stdin_config)
+    collection_rules = pc.get_conf_file()
+    rm_conf = pc.get_rm_conf()
     if rm_conf:
         logger.warn("WARNING: Excluding data from files")
 
@@ -366,6 +365,7 @@ def collect(config):
 
         archive = InsightsArchive(compressor=config.compressor,
                                   target_name=target['name'])
+        atexit.register(_delete_archive_internal, config, archive)
 
         # determine the target type and begin collection
         # we infer "docker_image" SPEC analysis for certain types
@@ -395,59 +395,38 @@ def collect(config):
 
 
 def get_connection(config):
-    global INSIGHTS_CONNECTION
-    if INSIGHTS_CONNECTION is None:
-        INSIGHTS_CONNECTION = InsightsConnection(config)
-    return INSIGHTS_CONNECTION
+    return InsightsConnection(config)
 
 
-def upload(config, tar_file, collection_duration=None):
+def _legacy_upload(config, pconn, tar_file, content_type, collection_duration=None):
     logger.info('Uploading Insights data.')
-    pconn = get_connection(config)
     api_response = None
     for tries in range(config.retries):
-        upload = pconn.upload_archive(tar_file, collection_duration)
+        upload = pconn.upload_archive(tar_file, '', collection_duration)
 
         if upload.status_code in (200, 201):
             api_response = json.loads(upload.text)
-            machine_id = generate_machine_id()
 
             # Write to last upload file
             with open(constants.last_upload_results_file, 'w') as handler:
-                handler.write(upload.text.encode('utf-8'))
+                if six.PY3:
+                    handler.write(upload.text)
+                else:
+                    handler.write(upload.text.encode('utf-8'))
             write_to_disk(constants.lastupload_file)
 
-            # Write to ansible facts directory
-            if os.path.isdir(constants.insights_ansible_facts_dir):
-                insights_facts = {}
-                insights_facts['last_upload'] = api_response
-
-                sat6 = _try_satellite6_configuration(config)
-                sat5 = None
-                if not sat6:
-                    sat5 = _try_satellite5_configuration(config)
-
-                if sat6:
-                    connection = 'sat6'
-                elif sat5:
-                    connection = 'sat5'
-                else:
-                    connection = 'rhsm'
-
-                insights_facts['conf'] = {'machine-id': machine_id, 'connection': connection}
-                with open(constants.insights_ansible_facts_file, 'w') as handler:
-                    handler.write(json.dumps(insights_facts))
-
+            msg_name = determine_hostname(config.display_name)
             account_number = config.account_number
             if account_number:
-                logger.info("Successfully uploaded report from %s to account %s." % (
-                            machine_id, account_number))
+                logger.info("Successfully uploaded report from %s to account %s.",
+                            msg_name, account_number)
             else:
-                logger.info("Successfully uploaded report for %s." % (machine_id))
+                logger.info("Successfully uploaded report for %s.", msg_name)
             break
 
         elif upload.status_code == 412:
             pconn.handle_fail_rcs(upload)
+            break
         else:
             logger.error("Upload attempt %d of %d failed! Status Code: %s",
                          tries + 1, config.retries, upload.status_code)
@@ -461,7 +440,40 @@ def upload(config, tar_file, collection_duration=None):
     return api_response
 
 
-def delete_archive(path):
+def upload(config, pconn, tar_file, content_type, collection_duration=None):
+    if config.legacy_upload:
+        return _legacy_upload(config, pconn, tar_file, content_type, collection_duration)
+    logger.info('Uploading Insights data.')
+
+    for tries in range(config.retries):
+        upload = pconn.upload_archive(tar_file, content_type, collection_duration)
+
+        if upload.status_code in (200, 202):
+            msg_name = determine_hostname(config.display_name)
+            logger.info("Successfully uploaded report for %s.", msg_name)
+        else:
+            logger.error("Upload attempt %d of %d failed! Status code: %s",
+                         tries + 1, config.retries, upload.status_code)
+            if tries + 1 != config.retries:
+                logger.info("Waiting %d seconds then retrying",
+                            constants.sleep_time)
+                time.sleep(constants.sleep_time)
+            else:
+                logger.error("All attempts to upload have failed!")
+                logger.error("Please see %s for additional information", config.logging_file)
+
+
+def _delete_archive_internal(config, archive):
+    '''
+    Only used during built-in collection.
+    Delete archive and tmp dirs on unexpected exit.
+    '''
+    if not config.keep_archive:
+        archive.delete_tmp_dir()
+        archive.delete_archive_file()
+
+
+def delete_archive(path, delete_parent_dir):
     removed_archive = False
 
     try:
@@ -471,11 +483,11 @@ def delete_archive(path):
         dirname = os.path.dirname
         abspath = os.path.abspath
         parent_tmp_dir = dirname(abspath(path))
-
-        logger.debug("Detected parent temporary directory %s", parent_tmp_dir)
-        if parent_tmp_dir != "/var/tmp" and parent_tmp_dir != "/var/tmp/":
-            logger.debug("Removing %s", parent_tmp_dir)
-            shutil.rmtree(parent_tmp_dir)
+        if delete_parent_dir:
+            logger.debug("Detected parent temporary directory %s", parent_tmp_dir)
+            if parent_tmp_dir != "/var/tmp" and parent_tmp_dir != "/var/tmp/":
+                logger.debug("Removing %s", parent_tmp_dir)
+                shutil.rmtree(parent_tmp_dir)
 
     except:
         logger.error("Error removing %s", path)
