+++ insights/client/utilities.py
@@ -10,7 +10,6 @@ import datetime
 import shlex
 import re
 import stat
-import json
 from subprocess import Popen, PIPE, STDOUT
 from six.moves.configparser import RawConfigParser
 
@@ -95,6 +94,8 @@ def write_to_disk(filename, delete=False, content=get_time()):
     """
     Write filename out to disk
     """
+    if not os.path.exists(os.path.dirname(filename)):
+        return
     if delete:
         if os.path.lexists(filename):
             os.remove(filename)
@@ -122,14 +123,6 @@ def generate_machine_id(new=False,
         logger.debug("Creating %s", destination_file)
         write_to_disk(destination_file, content=machine_id)
 
-    # update the ansible machine id facts file
-    if os.path.isdir(constants.insights_ansible_facts_dir):
-        if not (os.path.isfile(constants.insights_ansible_machine_id_file) and machine_id) or new:
-            machine_id_json = {'machine-id': machine_id}
-            with open(constants.insights_ansible_machine_id_file, 'w') as handler:
-                logger.debug('Writing Ansible machine-id facts file %s', constants.insights_ansible_machine_id_file)
-                handler.write(json.dumps(machine_id_json))
-
     return str(machine_id).strip()
 
 
@@ -153,7 +146,7 @@ def _expand_paths(path):
         logger.debug("Could not expand %s", path)
 
 
-def validate_remove_file(remove_file=constants.collection_remove_file):
+def validate_remove_file(remove_file):
     """
     Validate the remove file
     """
@@ -228,3 +221,45 @@ def modify_config_file(updates):
     cmd = cmd + constants.default_conf_file
     status = run_command_get_output(cmd)
     write_to_disk(constants.default_conf_file, content=status['output'])
+
+
+def get_version_info():
+    '''
+    Get the insights client and core versions for archival
+    '''
+    from insights.client import InsightsClient
+
+    cmd = 'rpm -q --qf "%{VERSION}-%{RELEASE}" insights-client'
+    version_info = {}
+    version_info['core_version'] = InsightsClient().version()
+    version_info['client_version'] = run_command_get_output(cmd)['output']
+
+    return version_info
+
+
+def print_egg_versions():
+    '''
+    Log all available eggs' versions
+    '''
+    versions = get_version_info()
+    logger.debug('Client version: %s', versions['client_version'])
+    logger.debug('Core version: %s', versions['core_version'])
+    logger.debug('All egg versions:')
+    eggs = [
+        os.getenv('EGG'),
+        '/var/lib/insights/newest.egg',
+        '/var/lib/insights/last_stable.egg',
+        '/etc/insights-client/rpm.egg',
+    ]
+    for egg in eggs:
+        if egg is None:
+            logger.debug('ENV egg not defined.')
+            continue
+        if not os.path.exists(egg):
+            logger.debug('%s not found.', egg)
+            continue
+        proc = Popen(['python', '-c', 'from insights.client import InsightsClient; print(InsightsClient().version())'],
+                     env={'PYTHONPATH': egg, 'PATH': os.getenv('PATH')}, stdout=PIPE, stderr=STDOUT)
+        stdout, stderr = proc.communicate()
+        version = stdout.decode('utf-8', 'ignore').strip()
+        logger.debug('%s: %s', egg, version)
